Fixed up the the example code
Fixed up the example code. Still two bugs though. Gets a display error
on the add screen because it is trying to get the 'models' list, which
is not sent to the add page. Also the change button actually adds a new
one, since there is no place in the example code that has an update sql
statement.