Can again interact with other skins besides skin 1
Buttons on show_champion.php are now part of a form, which submits the
champion id and the skin id. This means that we can again collect and
wish for skins other than skin 1.

Introduced a new bug though, now if you add a skin to the wishlist and
then add it to the collection this will result in 2 records in the
collection table. Needs to be fixed by giving the update button a
different value.